apps/bedelia/requirements.txt
# =============================
# App_Bedelia – Dependencies
# =============================
python

# Web framework
Flask==3.0.0

# MongoDB client (Replica Set support)
pymongo==4.6.1

# Redis client
redis==5.0.1

# MQTT client (TLS support)
paho-mqtt==1.6.1

# Environment variables
python-dotenv==1.0.0

# Password hashing
bcrypt==4.1.2

# Logging utils
colorlog==6.8.2
